+++ SandboxiePlus/SandMan/Windows/OptionsTemplates.cpp
@@ -257,10 +257,10 @@ void COptionsWindow::LoadFolders()
 	{
 		QSharedPointer<CSbieIni> pTemplate = QSharedPointer<CSbieIni>(new CSbieIni("Template_" + Name, m_pBox->GetAPI()));
 
-		QList<QPair<QString, QString>> AllValues = pTemplate->GetIniSection(NULL, true);
-		for (QList<QPair<QString, QString>>::const_iterator I = AllValues.begin(); I != AllValues.end(); ++I)
+		QList<CSbieIni::SbieIniValue> AllValues = pTemplate->GetIniSection(NULL, true);
+		for (QList<CSbieIni::SbieIniValue>::const_iterator I = AllValues.begin(); I != AllValues.end(); ++I)
 		{
-			QString Value = I->second;
+			QString Value = I->Value;
 
 			int begin = Value.indexOf("%Tmpl.");
 			if (begin == -1) continue;